+++ app/src/main/java/tarsila/costalonga/notasapp/ui/about/AboutFragment.kt
@@ -1,4 +1,4 @@
-package tarsila.costalonga.notasapp.ui.aboutsection
+package tarsila.costalonga.notasapp.ui.about
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
-import tarsila.costalonga.notasapp.ui.aboutsection.compose.AboutScreen
+import tarsila.costalonga.notasapp.ui.about.compose.AboutScreen
 import tarsila.costalonga.notasapp.ui.core.compose.theme.NotaComposeTheme
 
 class AboutFragment : Fragment() {
